pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "prometheus-client>=0.20.0",
    "psutil>=5.9.0",
    "asyncio-mqtt>=0.16.0",
]
